stt-tracker/requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
httpx>=0.27.0
